# Family-Nutrition-Planner/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NutriNext / ParivarSehat AI — Family Nutrition Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice/transcribe:
    post:
      operationId: transcribeVoice
      tags: [voice]
      summary: Transcribe audio using Sarvam AI (Hindi/English multilingual)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeVoiceBody"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeVoiceResponse"
  /voice/parse-profile:
    post:
      operationId: parseVoiceProfile
      tags: [voice]
      summary: Parse voice transcript into structured family profile fields using Gemini AI
      description: |
        Takes a Sarvam AI transcription (Hindi/English/regional) and uses Gemini to extract
        structured family data: name, state, budget, members with roles/ages/health conditions.
        This enables voice-driven onboarding for rural users.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseVoiceProfileBody"
      responses:
        "200":
          description: Structured family profile extracted from voice transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedVoiceProfile"

  # ── Demo ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TranscribeVoiceBody:
      type: object
      properties:
        audioBase64:
          type: string
          description: Base64 encoded audio (WAV/MP3)
        languageCode:
          type: string
          description: "BCP-47 code: hi-IN, en-IN, bn-IN, ta-IN, te-IN"
      required:
        - audioBase64
        - languageCode
    TranscribeVoiceResponse:
      type: object
      properties:
        transcript:
          type: string
        confidence:
          type: number
        detectedLanguage:
          type: string
      required:
        - transcript
        - confidence

    # ── Demo ───────────────────────────────────────────────────────────────────
...
    ParseVoiceProfileBody:
      type: object
      properties:
        transcript:
          type: string
          minLength: 1
        language:
          type: string
          enum: [hindi, english, bengali, tamil, telugu]
      required:
        - transcript
    ParsedVoiceMember:
      type: object
      properties:
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [father, mother, child, grandparent, other]
        age:
          type: integer
          nullable: true
        gender:
          type: string
          enum: [male, female, other]
        healthConditions:
          type: array
          items:
            type: string
        dietaryRestrictions:
          type: array
          items:
            type: string
      required:
        - role
        - gender
        - healthConditions
        - dietaryRestrictions
    ParsedVoiceProfile:
      type: object
      properties:
        familyName:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        monthlyBudget:
          type: number
          nullable: true
        primaryLanguage:
          type: string
          nullable: true
        cuisinePreferences:
          type: array
          items:
            type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/ParsedVoiceMember"
        confidence:
          type: number
          minimum: 0
          maximum: 1
        unparsedInfo:
          type: string
          nullable: true
      required:
        - cuisinePreferences
        - members
        - confidence

    # ── Food Scan (updated with confidence filtering) ─────────────────────────
...